+++ src/shared.c
@@ -473,7 +473,7 @@ int base32_encode (u8 (*f) (const u8), const u8 *in_buf, int in_len, u8 *out_buf
     out_ptr += 8;
   }
 
-  int out_len = (int) (((0.5 + (float) in_len) * 8) / 5); // ceil (in_len * 8 / 5)
+  int out_len = (int) (((0.5 + (double) in_len) * 8) / 5); // ceil (in_len * 8 / 5)
 
   while (out_len % 8)
   {
@@ -540,7 +540,7 @@ int base64_encode (u8 (*f) (const u8), const u8 *in_buf, int in_len, u8 *out_buf
     out_ptr += 4;
   }
 
-  int out_len = (int) (((0.5 + (float) in_len) * 8) / 6); // ceil (in_len * 8 / 6)
+  int out_len = (int) (((0.5 + (double) in_len) * 8) / 6); // ceil (in_len * 8 / 6)
 
   while (out_len % 4)
   {
@@ -5661,7 +5661,7 @@ void format_timer_display (struct tm *tm, char *buf, size_t len)
   }
 }
 
-void format_speed_display (float val, char *buf, size_t len)
+void format_speed_display (double val, char *buf, size_t len)
 {
   if (val <= 0)
   {
